feat: Add isProductDeleted field to Product model
This commit adds a new field `isProductDeleted` to the `Product` model in the `schema.prisma` file. The field is of type `Boolean` and has a default value of `false`. This addition enhances the functionality of the product management feature in the application by allowing products to be marked as deleted without actually removing them from the database.