feat: Add isProductDeleted field to Product schema
This commit adds the `isProductDeleted` field to the `Product` schema in the `admin-product-schema.ts` and `products-schema.ts` files. The field is of type `boolean` and has a default value of `false`. This addition enhances the functionality of the product management feature by allowing products to be marked as deleted without actually removing them from the database.