fix: Fix issue with ProductDeletedButton not closing modal after successful deletion
The ProductDeletedButton component was not closing the modal after successfully deleting a product. This commit fixes the issue by adding the setIsProductDeletedModalOpen(false) function call to the success branch of the deleteProductById function. Now, when a product is successfully deleted, the modal will be closed automatically, providing a better user experience.